% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrobiomeChangeMatrices.R
\name{calcMicrobiomeChanges}
\alias{calcMicrobiomeChanges}
\title{Calculates amount of quantitative, qualitative, or global change in microbiome between time points}
\usage{
calcMicrobiomeChanges(otus, changemeta, Ds = NULL, taxaAreRows)
}
\arguments{
\item{otus}{Matrix of taxon abundances (OTUs or ASVs). By default, rows are
samples (taxa are columns); this can be changed using the taxaAreRows argument.
Sample IDs should be appropriate dimension names and should match metadata. If
the matrix does not contain relative abundances, it will be converted to proportions
prior to calculating any volatility measures.}

\item{changemeta}{Data frame containing metadata, including columns for
the subject ID ("subjID"), time point 1 for the comparison ("time1"), time point 2 ("time2"),
sample ID corresponding to time point 1 ("sampID1"), and sample ID corresponding to
time point 2 ("sampID2"). The easiest way to generate this metadata file is through
the function temporalSubsampleMeta().}

\item{Ds}{Distance matrix or list of distance matrices, if distances between
comparison time points are desired. One column will be added to changemeta
for each distance matrix. Default NULL.}

\item{taxaAreRows}{Indicates whether taxa are rows (versus columns). Default FALSE.}
}
\value{
List containing the following elements:
\item{ChangeMeta}{New metadata for changes, with unique subject + time pair identifier and (if applicable) columns for distances }
\item{AddChangeMat}{Matrix with additive changes. Taxa are columns. Rows are labeled with new identifier.}
\item{MultChangeMat}{Matrix with multiplicative changes. Taxa are columns. Rows are labeled with new identifier.}
\item{QualChangeMat}{Matrix with qualitative changes. Taxa are columns. Rows are labeled with new identifier.}
\item{TaxCharacteristics}{Data frame with taxon IDs, average relative abundance, percentile of average relative abundance, and proportion of non-zero values across all samples. }
}
\description{
This function calculates several measures of change in the microbiome between
two consecutive time points based on relative abundance data, including:
\itemize{
\item Additive: t2-t1 (difference in taxon proportions. This is defined even when one or both proportions are zero.
\item Multiplicative: t2/t1 (ratio of taxon proportions). This will be NA for observations for which t1 = 0 and/or t2 = 0.
\item Qualitative: I(t2 > 0) - I(t1 > 0). This will be +1 if the taxon appeared (t2 > 0, t1 = 0); -1 if the taxon disappeared (t2 = 0, t1 > 0); and 0 otherwise.
It also returns a distances between consecutive time points if a distance matrix is provided by the user.
Finally, it calculates a few useful characteristics of the taxa themselves (average abundance across all samples, proportion of non-zero values).
}
}
\details{
Note that the function does NOT perform rarefaction prior to computing proportions.
The user will want to consider the possible impact of variable sampling depth
on results and account for it through rarefaction or another approach prior to
using these functions.
}
