% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gaj_meta}
\alias{gaj_meta}
\title{Gajer et al., Healthy Vaginal Microbiome metadata}
\format{
\subsection{\code{gaj_meta}}{

A data frame with 937 rows (samples) and 10 columns, including sampID (sample ID,
to match OTU table column names); subjID (subject ID); and time (in days).
Data access and preparation is described on GitHub at
https://github.com/aplantin/describing-volatility.
}
}
\source{
\url{https://doi.org/10.1126/scitranslmed.3003605}
}
\usage{
gaj_meta
}
\description{
Metadata associated with Gajer et al. (2012) OTU data.
}
\keyword{datasets}
