% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gaj_otu}
\alias{gaj_otu}
\title{Gajer et al., Healthy Vaginal Microbiome OTU table}
\format{
\subsection{\code{gaj_otu}}{

A data frame with 937 rows (samples) and 331 columns (taxa). Metadata
associated with sample IDs is included in gaj_meta. Data access and preparation
is described on GitHub at https://github.com/aplantin/describing-volatility.
}
}
\source{
\url{https://doi.org/10.1126/scitranslmed.3003605}
}
\usage{
gaj_otu
}
\description{
Gajer et al. (2012) sampled the vaginal microbiome of 32 healthy women
twice-weekly for 16 weeks, sequencing the V1-V2 region of the 16S rRNA gene.
Participants reported menstrual bleeding, sexual activity, medications and
contraceptives, and other characteristics in daily diaries. OTU counts were
computed based on the original study’s reported OTU proportions and total
read counts, rounded to the nearest whole number.
}
\keyword{datasets}
