% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mp_meta}
\alias{mp_meta}
\title{Moving Pictures gut timeseries metadata}
\format{
\subsection{\code{mp_meta}}{

A data frame with 467 rows (samples) and 7 columns:
\describe{
\item{sampID}{Sample ID, corresponding to column names of OTU table}
\item{age}{Subject age (years)}
\item{collectiontime}{Date of sample collection}
\item{subjID}{Subject ID}
\item{day}{Current day}
\item{firstday}{Earliest sampling date for that subject}
\item{time}{Number of days elapsed since first sampling date for that subject, calculated as day - firstday}
}
}
}
\source{
MicrobeDS R package
}
\usage{
mp_meta
}
\description{
Metadata associated with gut microbiome profiles for subjects M3 and F4 in
Caporaso et al.'s Moving Pictures study (all time points).
}
\keyword{datasets}
