% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mp_otu}
\alias{mp_otu}
\title{Moving Pictures gut timeseries OTU table}
\format{
\subsection{\code{mp_otu}}{

A data frame with 3,962 rows (taxa) and 467 columns (samples). Metadata
associated with sample IDs is included in mp_meta.
}
}
\source{
MicrobeDS R package
}
\usage{
mp_otu
}
\description{
Gut microbiome profiles for subjects M3 and F4 in Caporaso et al.'s Moving
Pictures study (all time points).
}
\keyword{datasets}
