% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mp_tax}
\alias{mp_tax}
\title{Moving Pictures taxonomic assignment for OTUs}
\format{
\subsection{\code{mp_tax}}{

A phylogenetic tree.
}
}
\source{
MicrobeDS R package
}
\usage{
mp_tax
}
\description{
Data frame showing taxonomic assignment for each OTU in Caporaso et al.'s
Moving Pictures study (all time points).
}
\keyword{datasets}
