% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mp_tree}
\alias{mp_tree}
\title{Moving Pictures phylogenetic tree}
\format{
\subsection{\code{mp_tree}}{

A phylogenetic tree.
}
}
\source{
MicrobeDS R package
}
\usage{
mp_tree
}
\description{
Phylogenetic tree associated with gut microbiome profiles for subjects M3 and F4 in
Caporaso et al.'s Moving Pictures study (all time points).
}
\keyword{datasets}
