% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rav_meta}
\alias{rav_meta}
\title{Ravel et al., Vaginal Microbiome (pre-BV)}
\format{
\subsection{\code{rav_meta}}{

A data frame with 186 rows (samples) and 21 columns, including sampID (sample ID,
to match OTU table column names); subjID (subject ID); and time (in days).
}
}
\source{
<https://microbiomejournal.biomedcentral.com/articles/10.1186/2049-2618-1-29 >
}
\usage{
rav_meta
}
\description{
Metadata associated with Ravel et al. (2013) OTU data. Data access and preparation is
described on GitHub at https://github.com/aplantin/describing-volatility.
}
\keyword{datasets}
