% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rav_otu}
\alias{rav_otu}
\title{Ravel et al., Vaginal Microbiome (pre-BV) OTU data}
\format{
\subsection{\code{rav_otu}}{

A data frame with 186 rows (samples) and 122 columns (taxa). Metadata
associated with sample IDs is included in rav_meta.
}
}
\source{
<https://microbiomejournal.biomedcentral.com/articles/10.1186/2049-2618-1-29 >
}
\usage{
rav_otu
}
\description{
Ravel et al. (2013) characterized the vaginal microbiome daily for 10 weeks
in 4 women without bacterial vaginosis (BV), 6 women with asymptomatic BV (ABV),
and 15 women with symptomatic BV (SBV), sequencing the V1-V3 regions of the
16S rRNA gene. For women with episodes of ABV or SBV, we include only
samples prior to the BV episode. Notably, women later diagnosed with ABV or
SBV typically have Lactobacillus-depleted vaginal microbiomes at earlier time
points, despite the lack of an active BV diagnosis based on Nugent score.
We exclude women with fewer than 20 remaining time points.
}
\keyword{datasets}
