% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RarefyExtensions.R
\name{rrarefy_variety}
\alias{rrarefy_variety}
\title{A variety of approaches to rarefaction}
\usage{
rrarefy_variety(
  otus,
  metadata,
  method = c("even_minreadct", "even_propminreadct", "subj_minreadct"),
  prop = NULL
)
}
\arguments{
\item{otus}{Matrix of taxon abundances (OTUs or ASVs). By default, rows are
samples (taxa are columns); this can be changed using the taxaAreRows argument.
Sample IDs should be appropriate dimension names and should match metadata.}

\item{metadata}{Data frame containing metadata, including columns containing
the subject ID ("subjID"), the sample ID ("sampID"), and time point.}

\item{method}{One of "even_minreadct" (all samples rarefied to the overall minimum
read count), "even_propminreadct" (all samples are rarefied to a constant proportion
of the minimum read count, specified in prop), and "subj_minreadct (all samples
for a subject are rarefied to that subject's minimum read count).}

\item{window_width}{Acceptable amount of "wiggle room" to each side of the
desired spacing. For example, if a 6-8 day range is acceptable for nominally
weekly sampling, use window_width = 1.}

\item{taxaAreRows}{Indicates whether taxa are rows (versus columns). Default FALSE.}

\item{seed}{}
}
\value{
New metadata file with columns subject ID, time point 1, time point 2, sample ID 1, and sample ID 2
}
\description{
This function cross-checks metadata samples with OTU samples, then rarefies
according to one of three approaches: (1) rarefying to the minimum read count
across all samples (the standard approach), (2) rarefying to some fraction of
the minimum read count (fraction specified in argument prop), and (3) rarefying
each subject's samples to that subject's minimum read count. It returns a
new OTU table with taxa as columns.
}
