% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smp_meta}
\alias{smp_meta}
\title{Student Microbiome Project gut timeseries metadata}
\format{
\subsection{\code{smp_meta}}{

A data frame with 496 rows (samples) and 17 columns, including sampID (sample ID,
to match OTU table column names); subjID (subject ID); and time (in days),
calculated as WeeksSinceStart * 7.
}
}
\source{
\url{https://github.com/caporaso-lab/student-microbiome-project}
}
\usage{
smp_meta
}
\description{
Metadata associated with gut microbiome profiles for 58 college-aged adults who
did not use antibiotics during the study period, taken from Flores et al.
(2014)'s Student Microbiome Project dataset.
}
\keyword{datasets}
