% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smp_otu}
\alias{smp_otu}
\title{Student Microbiome Project gut timeseries OTU table}
\format{
\subsection{\code{smp_otu}}{

A data frame with 10,509 rows (taxa) and 496 columns (samples). Metadata
associated with sample IDs is included in smp_meta.
}
}
\source{
\url{https://github.com/caporaso-lab/student-microbiome-project}
}
\usage{
smp_otu
}
\description{
Gut microbiome profiles for 58 college-aged adults who did not use antibiotics
during the study period, taken from Flores et al. (2014)'s Student Microbiome
Project dataset.
}
\keyword{datasets}
