% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smp_tax_key}
\alias{smp_tax_key}
\title{Student Microbiome Project gut timeseries metadata}
\format{
\subsection{\code{smp_tax_key}}{

A data frame with 1537 rows (genus-level taxa) and 2 columns: OTU ID (to match
dimension names of OTU table) and taxonomic assignment corresponding to that OTU ID.
}
}
\source{
\url{https://github.com/caporaso-lab/student-microbiome-project}
}
\usage{
smp_tax_key
}
\description{
Taxonomic assignment for OTUs in the gut microbiome profiles for 58 college-aged
adults who did not use antibiotics during the study period, taken from Flores et al.
(2014)'s Student Microbiome Project dataset.
}
\keyword{datasets}
