% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdHocVolatility.R
\name{summMicrobiomeVolatility}
\alias{summMicrobiomeVolatility}
\title{Summarizes measures of quantitative, qualitative, or global change in microbiome between time points}
\usage{
summMicrobiomeVolatility(mbchanges)
}
\arguments{
\item{mbchanges}{List of microbiome change matrices, obtained from calcMicrobiomeChanges()}
}
\value{
Data frame containing the following columns:
\item{AvgAbsAdditive}{Average absolute additive change in relative abundance (per taxon per subject).}
\item{AvgAbsAdditiveNZ}{Average absolute additive change in relative abundance \emph{between time points with nonzero abundances} (per taxon per subject)}
\item{AvgMultNZ}{Average multiplicative change in relative abundance \emph{between time points with nonzero abundances} (per taxon per subject)}
\item{AvgNZLogFC}{Average log fold change in relative abundance \emph{between time points with nonzero abundances} (per taxon per subject)}
\item{NumPairsNZ}{For each taxon for each subject, number of pairs of consecutive time points for which the taxon had nonzero abundance at both time points.}
\item{NumPairs}{Total number of pairs of time points for each subject (identical for all taxa)}
\item{PropQualChange}{Proportion of time points at which a taxon appeared or disappeared (per taxon per subject)}
\item{AvgTaxAbund}{Average relative abundance for the taxon (across all subjects and samples included in mbchanges)}
\item{PropTaxNZ}{Proportion of samples for which this taxon has nonzero abundance (across all subjects and samples included in mbchanges)}
}
\description{
This function calculates several ad hoc measures of microbiome volatility between
two consecutive time points based on changes in relative abundance (from calcMicrobiomeChanges()).
}
