% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TemporalSubsampling.R
\name{temporalSubsampleMeta}
\alias{temporalSubsampleMeta}
\title{Identifies time point subsamples}
\usage{
temporalSubsampleMeta(
  otus,
  metadata,
  desired_spacing,
  window_width,
  taxaAreRows = FALSE
)
}
\arguments{
\item{otus}{Matrix of taxon abundances (OTUs or ASVs). By default, rows are
samples (taxa are columns); this can be changed using the taxaAreRows argument.
Sample IDs should be appropriate dimension names and should match metadata.}

\item{metadata}{Data frame containing metadata, including columns containing
the subject ID ("subjID"), the sample ID ("sampID"), and time point.}

\item{desired_spacing}{Desired spacing between consecutive samples (in the
same units as the time column in metadata). For example, if time is measured
in days, then desired_spacing = 7 would give weekly sampling.}

\item{window_width}{Acceptable amount of "wiggle room" to each side of the
desired spacing. For example, if a 6-8 day range is acceptable for nominally
weekly sampling, use window_width = 1.}

\item{taxaAreRows}{Indicates whether taxa are rows (versus columns). Default FALSE.}
}
\value{
New metadata file with columns subject ID, time point 1, time point 2, sample ID 1, sample ID 2, and input parameters time lag and window width
}
\description{
This function cross-checks metadata samples with OTU samples, then identifies
pairs of samples separated by a fixed amount of time (with the ability to
specify an acceptable window around that fixed time difference). It returns a
new metadata file with the beginning and ending time point and sample ID for
each of the non-overlapping fixed-length windows.
}
